backend/requirements.txt
# UniLink Backend Dependencies
# Pin exact versions to prevent supply chain attacks

Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
python-dotenv==1.2.2
drf-spectacular==0.29.0
PyYAML==6.0.3
cryptography==46.0.6
PyJWT==2.12.1
requests==2.33.1
psycopg2-binary==2.9.11
dj-database-url==2.3.0

# Security audit
# Run: pip install pip-audit
# Run: pip-audit to check for vulnerabilities
